rename echoskillbot to skillbot
