+++ experimental/sso-with-skills/SSOWithSkills.sln
@@ -5,7 +5,7 @@ VisualStudioVersion = 16.0.29519.181
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SimpleRootBot", "SimpleRootBot\SimpleRootBot.csproj", "{966E5AB4-0697-4F8B-99CE-A80AE262A10A}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "EchoSkillBot", "EchoSkillBot\EchoSkillBot.csproj", "{5345ACF8-8695-4148-A8A5-44F335450297}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SkillBot", "SkillBot\SkillBot.csproj", "{5345ACF8-8695-4148-A8A5-44F335450297}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{EC72D0A6-71CD-4A94-A1D8-F30D92633514}"
 	ProjectSection(SolutionItems) = preProject
