+++ experimental/sso-with-skills/SimpleRootBot/Bots/RootBot.cs
@@ -50,7 +50,7 @@ namespace Microsoft.BotBuilderSamples.SimpleRootBot.Bots
             }
 
             // We use a single skill in this example.
-            var targetSkillId = "EchoSkillBot";
+            var targetSkillId = "SkillBot";
             if (!_skillsConfig.Skills.TryGetValue(targetSkillId, out _targetSkill))
             {
                 throw new ArgumentException($"Skill with ID \"{targetSkillId}\" not found in configuration");
