+++ experimental/sso-with-skills/SkillBot/Authentication/AllowedCallersClaimsValidator.cs
