+++ experimental/sso-with-skills/SkillBot/DeploymentTemplates/new-rg-parameters.json
