+++ experimental/sso-with-skills/SkillBot/DeploymentTemplates/preexisting-rg-parameters.json
