+++ experimental/sso-with-skills/SkillBot/DeploymentTemplates/template-with-new-rg.json
