+++ experimental/sso-with-skills/SkillBot/DeploymentTemplates/template-with-preexisting-rg.json
