+++ experimental/sso-with-skills/SkillBot/Dialogs/DisplayTokenDialog.cs
@@ -7,7 +7,7 @@ using Microsoft.Bot.Builder;
 using Microsoft.Bot.Builder.Dialogs;
 using Microsoft.Extensions.Configuration;
 
-namespace Microsoft.BotBuilderSamples.EchoSkillBot.Dialogs
+namespace Microsoft.BotBuilderSamples.SkillBot.Dialogs
 {
     public class DisplayTokenDialog : ComponentDialog
     {
