+++ experimental/sso-with-skills/SkillBot/README.md
@@ -1,3 +1,3 @@
-﻿# EchoSkillBot
+﻿# SkillBot
 
 See [SkillSimpleBotToBot](../) for details on how to configure and run this sample.
