+++ experimental/sso-with-skills/SkillBot/SkillBot.csproj
