+++ experimental/sso-with-skills/SkillBot/Startup.cs
@@ -8,14 +8,14 @@ using Microsoft.Bot.Builder.BotFramework;
 using Microsoft.Bot.Builder.Dialogs;
 using Microsoft.Bot.Builder.Integration.AspNet.Core;
 using Microsoft.Bot.Connector.Authentication;
-using Microsoft.BotBuilderSamples.EchoSkillBot.Authentication;
-using Microsoft.BotBuilderSamples.EchoSkillBot.Bots;
-using Microsoft.BotBuilderSamples.EchoSkillBot.Dialogs;
+using Microsoft.BotBuilderSamples.SkillBot.Authentication;
+using Microsoft.BotBuilderSamples.SkillBot.Bots;
+using Microsoft.BotBuilderSamples.SkillBot.Dialogs;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 
-namespace Microsoft.BotBuilderSamples.EchoSkillBot
+namespace Microsoft.BotBuilderSamples.SkillBot
 {
     public class Startup
     {
