+++ experimental/sso-with-skills/SkillBot/wwwroot/default.htm
@@ -4,7 +4,7 @@
 <head>
     <meta charset="utf-8">
     <meta name="viewport" content="width=device-width, initial-scale=1">
-    <title>EchoSkillBot</title>
+    <title>SkillBot</title>
     <style>
         body {
             margin: 0px;
@@ -392,7 +392,7 @@
     <header class="header">
         <div class="header-inner-container">
             <div class="header-icon" style="display: inline-block"></div>
-            <div class="header-text" style="display: inline-block">EchoSkillBot Bot</div>
+            <div class="header-text" style="display: inline-block">SkillBot Bot</div>
         </div>
     </header>
     <div class="row">
